Run API suites in scheduled workflow
