+++ starbase/ai-roomchat/__tests__/api/errors/report.test.js
@@ -1,53 +1,14 @@
 const { describe, it, expect, beforeEach, afterEach } = require('@jest/globals')
+const {
+  createApiRequest,
+  createMockResponse,
+  loadApiRoute,
+  registerSupabaseAdminMock,
+  createSupabaseInsertChain,
+} = require('../testUtils')
 
-function createMockResponse() {
-  const response = {
-    statusCode: 200,
-    headers: {},
-    body: null,
-    status(code) {
-      this.statusCode = code
-      return this
-    },
-    setHeader(name, value) {
-      this.headers[name] = value
-    },
-    json(payload) {
-      this.body = payload
-      return this
-    },
-  }
-  return response
-}
-
-function createRequest({
-  method = 'POST',
-  body = {},
-  headers = {},
-  ip,
-}) {
-  return {
-    method,
-    body,
-    headers,
-    socket: { remoteAddress: ip || headers['x-forwarded-for'] || '127.0.0.1' },
-  }
-}
-
-function mockSupabase(insertImplementation = () => Promise.resolve({ error: null })) {
-  const insertMock = jest.fn(insertImplementation)
-  const fromMock = jest.fn(() => ({ insert: insertMock }))
-  jest.doMock('@/lib/supabaseAdmin', () => ({
-    supabaseAdmin: {
-      from: fromMock,
-    },
-  }))
-  return { insertMock, fromMock }
-}
-
-async function loadHandler() {
-  const module = require('../pages/api/errors/report')
-  return module.default
+function loadHandler() {
+  return loadApiRoute('errors', 'report')
 }
 
 describe('POST /api/errors/report', () => {
@@ -61,9 +22,11 @@ describe('POST /api/errors/report', () => {
   })
 
   it('rejects non-POST methods', async () => {
-    mockSupabase()
-    const handler = await loadHandler()
-    const req = createRequest({ method: 'GET' })
+    const { fromMock } = createSupabaseInsertChain()
+    registerSupabaseAdminMock(fromMock)
+
+    const handler = loadHandler()
+    const req = createApiRequest({ method: 'GET' })
     const res = createMockResponse()
 
     await handler(req, res)
@@ -73,10 +36,12 @@ describe('POST /api/errors/report', () => {
   })
 
   it('validates presence of error message', async () => {
-    const { insertMock } = mockSupabase()
-    const handler = await loadHandler()
+    const { insertMock, fromMock } = createSupabaseInsertChain()
+    registerSupabaseAdminMock(fromMock)
 
-    const req = createRequest({ body: { message: '   ' } })
+    const handler = loadHandler()
+
+    const req = createApiRequest({ method: 'POST', body: { message: '   ' } })
     const res = createMockResponse()
 
     await handler(req, res)
@@ -87,11 +52,14 @@ describe('POST /api/errors/report', () => {
   })
 
   it('persists sanitized payloads', async () => {
-    const { insertMock } = mockSupabase()
-    const handler = await loadHandler()
+    const { insertMock, fromMock } = createSupabaseInsertChain()
+    registerSupabaseAdminMock(fromMock)
+
+    const handler = loadHandler()
 
     const longPath = '/dashboard'.padEnd(600, 'x')
-    const req = createRequest({
+    const req = createApiRequest({
+      method: 'POST',
       headers: {
         'user-agent': 'JestSuite/1.0',
         'x-forwarded-for': '203.0.113.1',
@@ -122,11 +90,14 @@ describe('POST /api/errors/report', () => {
   })
 
   it('enforces throttling per ip/session pair', async () => {
-    const { insertMock } = mockSupabase()
-    const handler = await loadHandler()
+    const { insertMock, fromMock } = createSupabaseInsertChain()
+    registerSupabaseAdminMock(fromMock)
+
+    const handler = loadHandler()
 
     for (let index = 0; index < 12; index += 1) {
-      const req = createRequest({
+      const req = createApiRequest({
+        method: 'POST',
         headers: { 'x-forwarded-for': '198.51.100.5' },
         body: { message: `Issue ${index}`, sessionId: 'abc' },
       })
@@ -135,7 +106,8 @@ describe('POST /api/errors/report', () => {
       expect(res.statusCode).toBe(201)
     }
 
-    const throttledReq = createRequest({
+    const throttledReq = createApiRequest({
+      method: 'POST',
       headers: { 'x-forwarded-for': '198.51.100.5' },
       body: { message: 'Final issue', sessionId: 'abc' },
     })
@@ -148,30 +120,18 @@ describe('POST /api/errors/report', () => {
   })
 
   it('returns 500 when persistence fails', async () => {
-    const { insertMock } = mockSupabase(() => Promise.resolve({ error: { message: 'db down' } }))
-    const handler = await loadHandler()
+    const { insertMock, fromMock } = createSupabaseInsertChain(() => Promise.resolve({ error: { message: 'db down' } }))
+    registerSupabaseAdminMock(fromMock)
 
-    const req = createRequest({ body: { message: 'boom' } })
+    const handler = loadHandler()
+
+    const req = createApiRequest({ method: 'POST', body: { message: 'boom' } })
     const res = createMockResponse()
 
     await handler(req, res)
 
     expect(res.statusCode).toBe(500)
     expect(res.body).toEqual({ error: 'Failed to persist error report' })
-    expect(insertMock).toHaveBeenCalledTimes(1)
-  })
-
-  it('handles unexpected insertion errors gracefully', async () => {
-    const { insertMock } = mockSupabase(() => Promise.reject(new Error('network failure')))
-    const handler = await loadHandler()
-
-    const req = createRequest({ body: { message: 'boom' } })
-    const res = createMockResponse()
-
-    await handler(req, res)
-
-    expect(res.statusCode).toBe(500)
-    expect(res.body).toEqual({ error: 'Unexpected error' })
     expect(insertMock).toHaveBeenCalledTimes(1)
   })
 })
