+++ starbase/ai-roomchat/docs/page-audit.md
@@ -15,7 +15,7 @@ This document captures a page-by-page walkthrough of the current Next.js `pages`
 
 ### `pages/index.js`
 - Renders a full-screen landing page with themed background, title (“랭크 청사진에 맞춘 전선으로 바로 합류하세요”), and `AuthButton` to kick off Supabase auth flow while auto-redirecting authenticated users to `/roster`.【F:pages/index.js†L1-L118】
-- Uses `Home.module.css` for glassmorphism styling, a core feature list, and blueprint overview widgets that now pull priority/리소스 메타 alongside 담당자·기한 정보를 JSON에서 읽어와 D-Day 배지, 우선순위/예상 리소스 칩, 정렬 토글(우선순위·기한·목록순)을 함께 노출합니다. `npm run refresh:blueprint-progress` regenerates the JSON payloads and 문서 경고 블록 from the overview doc so the UI stays aligned without manual edits, `npm run check:blueprint-progress-freshness` (also executed weekly via GitHub Actions) fails when the snapshot is older than the 14일 한계치, and `npm run check:blueprint-next-actions` fails CI when 마감이 지난 항목이 발견돼 문서/랜딩 모두에서 경고를 즉시 띄울 수 있습니다.【F:pages/index.js†L52-L246】【F:styles/Home.module.css†L1-L332】【F:data/rankBlueprintProgress.json†L1-L35】【F:data/rankBlueprintNextActions.json†L1-L60】【F:scripts/refresh-rank-blueprint-progress.js†L1-L362】【F:scripts/check-rank-blueprint-progress-freshness.js†L1-L74】【F:scripts/check-rank-blueprint-next-actions.js†L1-L87】【F:.github/workflows/blueprint-progress-freshness.yml†L1-L22】
+- Uses `Home.module.css` for glassmorphism styling, a core feature list, and blueprint overview widgets that now pull priority/리소스 메타 alongside 담당자·기한 정보를 JSON에서 읽어와 D-Day 배지, 우선순위/예상 리소스 칩, 정렬 토글(우선순위·기한·목록순)을 함께 노출합니다. `npm run refresh:blueprint-progress` regenerates the JSON payloads and 문서 경고 블록 from the overview doc so the UI stays aligned without manual edits, `npm run check:blueprint-progress-freshness` (also executed weekly via GitHub Actions) fails when the snapshot is older than the 14일 한계치, `npm run check:blueprint-next-actions` fails CI when 마감이 지난 항목이 발견돼 문서/랜딩 모두에서 경고를 즉시 띄울 수 있고, 동일 워크플로가 `npm test -- --runInBand`와 `CI=1 npm run build`까지 실행해 에러 리포트 API 회귀를 빠르게 감지합니다.【F:pages/index.js†L52-L246】【F:styles/Home.module.css†L1-L332】【F:data/rankBlueprintProgress.json†L1-L35】【F:data/rankBlueprintNextActions.json†L1-L60】【F:scripts/refresh-rank-blueprint-progress.js†L1-L362】【F:scripts/check-rank-blueprint-progress-freshness.js†L1-L74】【F:scripts/check-rank-blueprint-next-actions.js†L1-L87】【F:.github/workflows/blueprint-progress-freshness.yml†L1-L33】
 
 ### `pages/auth-callback.js`
 - Client-side handler for Supabase OAuth redirects. Parses the current URL, forwards to `handleOAuthCallback`, and responds based on returned status (direct redirect vs. error with retry).【F:pages/auth-callback.js†L1-L35】
@@ -105,6 +105,7 @@ This document captures a page-by-page walkthrough of the current Next.js `pages`
 
 ## Additional Observations & Potential Follow-Ups
 - 클라이언트 전역 오류 리포터(`ClientErrorReporter`)가 `/api/errors/report`로 전송한 스냅샷을 관리자 포털의 `UserErrorMonitor`가 `/api/admin/errors`를 통해 집계해 운영팀이 현장 오류를 즉시 파악할 수 있습니다.【F:components/ClientErrorReporter.js†L1-L119】【F:components/admin/UserErrorMonitor.js†L1-L94】
+- `/api/errors/report`와 `/api/admin/errors` 라우트는 `__tests__/api` 스위트와 헬퍼(`testUtils`)로 커버돼 있어 향후 관리자 API가 늘어나도 동일 패턴으로 테스트를 복제하며 회귀를 방지할 수 있습니다.【F:__tests__/api/testUtils.js†L1-L82】【F:__tests__/api/errors/report.test.js†L1-L137】【F:__tests__/api/admin/errors.test.js†L1-L183】
 - Supabase table helpers (`withTable`) abstract table name prefix differences across environments—most hooks and API routes rely on them for multi-tenant support.
 - Several UI shells (create, roster, maker) defer most logic to component containers; reviewing those components is recommended for full domain context beyond this page-oriented audit.
 - Ensure environment variables (`NEXT_PUBLIC_SUPABASE_URL`, `SUPABASE_SERVICE_ROLE`, etc.) are set for server APIs like `finalize-session` to work outside local mocks.【F:pages/api/rank/finalize-session.js†L1-L8】
