+++ starbase/ai-roomchat/docs/rank-blueprint-progress-2025-11-07.md
@@ -58,7 +58,7 @@
 - 내보내기와 동시에 `/api/rank/upload-cooldown-timeline`이 팀 드라이브 업로드를 수행해 Slack 공유 없이도 운영팀이 최신 보고 파일을 열람할 수 있습니다.
 - `/api/rank/upload-cooldown-timeline` 요청이 업로드 결과를 `rank_cooldown_timeline_uploads`에 기록하도록 하고, 텔레메트리 `timelineUploads` 요약과 대시보드 업로드 현황 패널을 추가해 누락된 공유 구간을 바로 파악할 수 있습니다.
 - 업로드 연속 실패/장기 미성공 임계값을 `cooldownAlertThresholds`와 텔레메트리 경보(`alerts.timelineUploads`)에 연결해 카드·이슈 패널·CSV 내보내기에서도 동일한 경고를 표시하도록 정비했습니다.
-- 랜딩 히어로에 청사진 주요 단계 진행률(매칭 트리거 통일, 세션/전투 동기화, 프롬프트 변수 자동화, UI·오디오 완성, 운영 가드)을 2025-11-07 기준으로 노출하는 보드를 추가하고, 진행률 데이터를 `data/rankBlueprintProgress.json`으로 분리해 업데이트 시 문서/코드 동기화를 간소화했습니다. `npm run refresh:blueprint-progress` 스크립트가 개요 문서의 표를 읽어 JSON을 재생성하도록 연결해 후속 갱신도 한 번에 처리할 수 있고, `npm run check:blueprint-progress-freshness` + 주 1회 GitHub Actions 리마인더로 14일이 지나기 전에 자동 경고를 받습니다.
+- 랜딩 히어로에 청사진 주요 단계 진행률(매칭 트리거 통일, 세션/전투 동기화, 프롬프트 변수 자동화, UI·오디오 완성, 운영 가드)을 2025-11-07 기준으로 노출하는 보드를 추가하고, 진행률 데이터를 `data/rankBlueprintProgress.json`으로 분리해 업데이트 시 문서/코드 동기화를 간소화했습니다. `npm run refresh:blueprint-progress` 스크립트가 개요 문서의 표를 읽어 JSON을 재생성하도록 연결해 후속 갱신도 한 번에 처리할 수 있고, `npm run check:blueprint-progress-freshness` + 주 1회 GitHub Actions 리마인더로 14일이 지나기 전에 자동 경고를 받습니다. 같은 워크플로가 `npm test -- --runInBand`와 `CI=1 npm run build`도 수행해 에러 리포트 파이프라인 회귀를 즉시 감지합니다.
 - 같은 흐름으로 "다음 액션 스냅샷" 섹션도 `data/rankBlueprintNextActions.json`으로 추출해 랜딩 히어로에 진행률 보드 옆 체크리스트 카드를 노출했습니다. 담당자와 목표일을 표 형태로 관리해 JSON에도 `owner`·`targetDateISO` 메타가 포함되고, 랜딩 카드에서 D-Day 뱃지·컬러 상태로 우선순위가 드러나도록 구성했습니다. 최신 업데이트 시각은 진행률 카드와 동일한 뱃지를 공유하며, `npm run refresh:blueprint-progress`를 실행하면 두 JSON이 동시에 갱신됩니다.
 - 자동 경고 흐름을 확장해 다음 액션 표에서 담당자 키와 목표일 메타를 파싱하면서 JSON에 기한 상태(D-Day, 연체 여부, 남은 일수)를 기록하고, 개요 문서에는 경고 블록을 자동 삽입하도록 스크립트를 보강했습니다. 주간 GitHub Actions는 `npm run check:blueprint-next-actions`를 함께 실행해 마감이 지나면 즉시 실패하도록 구성했습니다.
 - 다음 액션 JSON에 `priority`·`effort` 메타를 추출하고 랜딩 카드에 우선순위 뱃지, 예상 리소스 칩, 정렬 토글(우선순위/기한/목록순)을 추가해 청사진 표와 동일 기준으로 우선순위와 공수 정보를 비교할 수 있게 했습니다.
